# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.CreateCompanyRequest:
    properties:
      name:
        example: Tech Solutions LLC
        maxLength: 200
        minLength: 2
        type: string
      password:
        example: company123
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "+998901112233"
        maxLength: 30
        minLength: 9
        type: string
      username:
        example: techsol
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - password
    - phone
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse:
    properties:
      error:
        example: validation failed
        type: string
      message:
        example: Xatolik yuz berdi
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.LoginRequest:
    properties:
      password:
        example: company123
        type: string
      username:
        example: techsol
        type: string
    required:
    - password
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse:
    properties:
      data: {}
      message:
        example: Muvaffaqiyatli
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.UpdateCompanyRequest:
    properties:
      name:
        example: Tech Solutions LLC
        maxLength: 200
        minLength: 2
        type: string
      password:
        example: newcompany123
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "+998901112233"
        maxLength: 30
        minLength: 9
        type: string
      username:
        example: techsol
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - phone
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.UpdateProfileRequest:
    properties:
      name:
        example: Tech Solutions LLC
        maxLength: 200
        minLength: 2
        type: string
      password:
        example: newcompany123
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "+998901112233"
        maxLength: 30
        minLength: 9
        type: string
      username:
        example: techsol
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    - phone
    - username
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse:
    properties:
      error:
        example: validation failed
        type: string
      message:
        example: Xatolik yuz berdi
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.MessageResponse:
    properties:
      data: {}
      message:
        example: Muvaffaqiyatli
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.UpdateSurveyLinkBaseRequest:
    properties:
      base_url:
        example: http://localhost:3000/s
        maxLength: 500
        minLength: 8
        type: string
    required:
    - base_url
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Modular monolit backend. Admin, Company, Survey va Settings modullari.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Foydalanuvchilar bazasi API
  version: "1.0"
paths:
...
  /companies:
    get:
      description: Barcha kompaniyalarni qaytaradi (admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kompaniyalar ro'yxati
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Admin tomonidan yangi kompaniya qo'shiladi
      parameters:
      - description: Kompaniya ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Yangi kompaniya yaratish
      tags:
      - Companies
  /companies/{id}:
    delete:
      description: Kompaniyani tizimdan o'chiradi (admin)
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kompaniyani o'chirish
      tags:
      - Companies
    get:
      description: Bitta kompaniya ma'lumotlarini qaytaradi (admin)
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kompaniyani ID bo'yicha olish
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Kompaniya ma'lumotlarini o'zgartiradi. Password bo'sh qoldirilsa,
        eski parol saqlanadi.
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      - description: Yangilangan ma'lumotlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kompaniyani yangilash
      tags:
      - Companies
  /company/auth/login:
    post:
      consumes:
      - application/json
      description: Username va password orqali kompaniya tizimga kiradi. JWT token
        qaytaradi.
      parameters:
      - description: Login ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      summary: Kompaniya login
      tags:
      - Company Auth
  /company/auth/profile:
    get:
      description: Token egasi kompaniyaning profil ma'lumotlarini qaytaradi
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Joriy kompaniya profili
      tags:
      - Company Auth
    put:
      consumes:
      - application/json
      description: Token egasi o'z profilini o'zgartiradi. Password bo'sh qoldirilsa,
        eski parol saqlanadi.
      parameters:
      - description: Profil ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_company_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Joriy kompaniya profilini yangilash
      tags:
      - Company Auth
...
  /settings/survey-link-base:
    get:
      description: So'rovnomalar uchun public linkning asosiy (base) URL manzilini
        qaytaradi
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: So'rovnoma link base URL olish
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Admin so'rovnoma public linklarining asosiy URL manzilini o'zgartiradi.
        Mavjud so'rovnomalar linklari yangi base bo'yicha avtomatik hisoblanadi.
      parameters:
      - description: Base URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.UpdateSurveyLinkBaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_setting_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: So'rovnoma link base URL yangilash
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"